pyproject.toml
[project]
name = "oldnews"
version = "0.8.0"
description = "A terminal-based client for TheOldReader"
readme = "README.md"
authors = [
    { name = "Dave Pearson", email = "davep@davep.org" }
]
requires-python = ">=3.12"
license = "GPL-3.0-or-later"
keywords = [
    "atom",
    "client",
    "Google Reader",
    "RSS",
    "TheOldReader",
    "terminal",
    "news-reader",
    "news",
]
dependencies = [
    "bagofstuff>=0.1.0",
    "html-to-markdown>=2.15.0",
    "oldas>=0.10.0",
    "textual>=6.3.0",
    "textual-autocomplete>=4.0.6",
    "textual-enhanced>=1.2.0",
    "tortoise-orm>=0.25.4",
    "typedal>=4.2.2",
    "xdg-base-dirs>=6.0.2",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Typing :: Typed",
]

[project.urls]
Homepage = "https://github.com/davep/oldnews"
Repository = "https://github.com/davep/oldnews"
Documentation = "https://oldnews.davep.dev/"
Source = "https://github.com/davep/oldnews"
Issues = "https://github.com/davep/oldnews/issues"
Discussions = "https://github.com/davep/oldnews/discussions"

[project.scripts]
oldnews = "oldnews.__main__:main"

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true

[tool.pyright]
venvPath="."
venv=".venv"
exclude=[".venv"]

[dependency-groups]
dev = [
    "codespell>=2.4.1",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "ruff>=0.14.0",
    "textual-dev>=1.8.0",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
ignore = [
    # I think try...expect...pass reads far better.
    "SIM105",
]

[tool.ruff.lint.pycodestyle]
max-line-length = 120
